remove panels
