+++ packages/databricks-vscode/src/cluster/ClusterCommands.ts
@@ -1,32 +1,10 @@
 import {ConnectionManager} from "../configuration/ConnectionManager";
-import {ClusterFilter, ClusterModel} from "./ClusterModel";
 
 /**
  * Cluster related commands
  */
 export class ClusterCommands {
-    constructor(
-        private clusterModel: ClusterModel,
-        readonly connectionManager: ConnectionManager
-    ) {}
-
-    /**
-     * Refresh cluster tree view by reloading them throug the API
-     */
-    refreshCommand() {
-        return () => {
-            this.clusterModel.refresh();
-        };
-    }
-
-    /**
-     * Command to filter clusters in the cluster tree view
-     */
-    filterCommand(filter: ClusterFilter) {
-        return () => {
-            this.clusterModel.filter = filter;
-        };
-    }
+    constructor(readonly connectionManager: ConnectionManager) {}
 
     /**
      * Command to start a cluster
