+++ packages/databricks-vscode/src/workspace-fs/index.ts
@@ -1,2 +1,1 @@
-export * from "./WorkspaceFsDataProvider";
 export * from "./WorkspaceFsCommands";
